requirements.txt
Django==3.0.4

dj-database-url==0.5.0
djangorestframework==3.11.0
psycopg2==2.8.4
python-dotenv==0.12.0
